+++ tests/testthat/roxygen-examples-complete/17-two-no-non-comment-out.R
@@ -4,6 +4,7 @@
 #' }
 #' # before this comment is a left-over space
 #' another_function <- function() NULL
+#'
 #' @examples
 #'   my_fun <- function() {
 #'   print("hello world!")