+++ tests/testthat/roxygen-examples-complete/26-empty-trailing-lines-in.R
@@ -1,0 +1,30 @@
+#' this
+#'
+#' mey
+#' @examples
+#' 2 + 1
+#'
+NULL
+
+
+#' this
+#'
+#' mey
+#' @examples
+#' 2 + 1
+#'
+#'
+#'
+#'
+#'
+#'
+#'
+NULL
+
+
+#' this
+#'
+#' mey
+#' @examples
+#' 2 + 1
+NULL